+++ frontend/src/cards/ui/TerritoryCircles.tsx
@@ -3,7 +3,8 @@ import {
   type DataTypeConfig,
   type MetricConfig,
 } from '../../data/config/MetricConfig'
-import { Fips, TERRITORY_CODES } from '../../data/utils/Fips'
+import { Fips } from '../../data/utils/Fips'
+
 import { type DemographicGroup } from '../../data/utils/Constants'
 import { type Row } from '../../data/utils/DatasetTypes'
 import { type DemographicType } from '../../data/query/Breakdowns'
@@ -12,6 +13,7 @@ import {
   type CountColsMap,
   unknownMapConfig,
 } from '../../charts/mapGlobals'
+import { TERRITORY_CODES } from '../../data/utils/ConstantsGeography'
 
 interface TerritoryCirclesProps {
   data: Array<Record<string, any>>
