+++ frontend/src/styles/HetComponents/HetBreadcrumbs.tsx
@@ -1,8 +1,9 @@
-import { USA_FIPS, USA_DISPLAY_NAME, Fips } from '../../data/utils/Fips'
 import Breadcrumbs from '@mui/material/Breadcrumbs'
 import Button from '@mui/material/Button'
 import { useLocation } from 'react-router-dom'
 import { type ScrollableHashId } from '../../utils/hooks/useStepObserver'
+import { USA_DISPLAY_NAME, USA_FIPS } from '../../data/utils/ConstantsGeography'
+import { Fips } from '../../data/utils/Fips'
 
 export default function HetBreadcrumbs(props: {
   fips: Fips
