+++ frontend/src/utils/MadLibs.ts
@@ -15,7 +15,8 @@ import {
 } from '../data/config/MetricConfigPhrma'
 import { SDOH_CATEGORY_DROPDOWNIDS } from '../data/config/MetricConfigSDOH'
 import { SHOW_PHRMA_MENTAL_HEALTH } from '../data/providers/PhrmaProvider'
-import { FIPS_MAP, GEORGIA_FIPS, USA_FIPS } from '../data/utils/Fips'
+import { GEORGIA_FIPS, USA_FIPS } from '../data/utils/ConstantsGeography'
+import { FIPS_MAP } from '../data/utils/FipsData'
 
 // Map of phrase segment index to its selected value
 export type PhraseSelections = Record<number, string>
